Improve performance of argument propagation
Argument propagation performs a top-down propagation over the class
hierarchy, where it pushes argument information ("method states") for
virtual methods downwards to overrides.

Some of this argument information only becomes active when it reaches a
given class in the class hierarchy. Currently, this is implemented by
copying these method states downwards in the class hierarchy traversal
and activating them when when the "activating class type" is met.

This CL changes the implementation to move such argument information
directly to the activating class type, which significantly improves
performance in some cases.

Bug: b/190154391
Change-Id: Ie23e985b4addb3534c34306ba699872f20738ca7